feat(debug): Add Tunables Tab (#3431)
Added a Tunables tab similar to Globals/Locals that can be used to read and write tunables by inputting their unsigned/signed or hexadecimal hash, as well as their pre-joaat name. It also displays the offset of the given tunable.

Removed the animation players in the Debug tab as the one in the Self tab makes them redundant.